+++ packages/ethereum-provider/src/index.ts
@@ -1,17 +1,13 @@
-import { Address, Binary, Word } from "@rarible/types"
-import { SignTypedDataMethodEnum, TypedSignatureData } from "./domain"
+import type { Address, Binary, Word } from "@rarible/types"
+import type { MessageTypes, TypedMessage } from "./domain"
+import { signTypedDataInternal } from "./utils"
 
 export interface EthereumTransaction {
 	hash: Word
-
 	from: Address
-
 	to?: Address
-
 	data: Binary
-
 	nonce: number
-
 	wait(): Promise<void>
 }
 
@@ -23,7 +19,6 @@ export interface EthereumSendOptions {
 
 export interface EthereumFunctionCall {
 	call(options?: EthereumSendOptions): Promise<any>
-
 	send(options?: EthereumSendOptions): Promise<EthereumTransaction>
 }
 
@@ -33,28 +28,14 @@ export interface EthereumContract {
 
 export interface Ethereum {
 	createContract(abi: any, address?: string): EthereumContract
-
 	send(method: string, params: any): Promise<any>
-
 	getFrom(): Promise<string>
-
 	personalSign(message: string): Promise<string>
-
 	sha3(string: string): string
-
 	encodeParameter(type: any, parameter: any): string
 }
 
-
-export async function signTypedData(ethereum: Ethereum, data: TypedSignatureData) {
+export async function signTypedData<T extends MessageTypes>(ethereum: Ethereum, data: TypedMessage<T>) {
 	const signer = await ethereum.getFrom()
-	try {
-		return await ethereum.send(SignTypedDataMethodEnum.V4, [signer, JSON.stringify(data)])
-	} catch (error) {
-		try {
-			return await ethereum.send(SignTypedDataMethodEnum.V3, [signer, JSON.stringify(data)])
-		} catch (error) {
-			return await ethereum.send(SignTypedDataMethodEnum.DEFAULT, [signer, data])
-		}
-	}
+	return signTypedDataInternal(signer, ethereum, data)
 }