+++ packages/protocol-ethereum-sdk/src/common/send-transaction.test.ts
@@ -41,5 +41,5 @@ describe("sendTransaction", () => {
 		const logs = await createPendingLogs(gatewayApi, tx)
 		expect(logs).toBeTruthy()
 		expect(tx.from.toLowerCase()).toBe(minter.toLowerCase())
-	})
+	}, 10000)
 })