+++ packages/protocol-ethereum-sdk/src/nft/sign-nft.ts
@@ -1,15 +1,15 @@
-import { Binary, EIP712Domain, LazyErc1155 } from "@rarible/protocol-api-client"
+import type { Binary, EIP712Domain, LazyErc1155, LazyErc721 } from "@rarible/protocol-api-client"
 import { Address, toBinary } from "@rarible/types"
 import { signTypedData } from "@rarible/ethereum-provider"
-import { LazyErc721 } from "@rarible/protocol-api-client/build/models/LazyNft"
-import { Ethereum } from "@rarible/ethereum-provider"
+import type { Ethereum } from "@rarible/ethereum-provider/build"
+import type { TypedMessage } from "eth-sig-util"
 import {
 	EIP1155_DOMAIN_NFT_TEMPLATE,
 	EIP1155_NFT_TYPE,
 	EIP1155_NFT_TYPES,
 	EIP721_DOMAIN_NFT_TEMPLATE,
 	EIP721_NFT_TYPE,
-	EIP721_NFT_TYPES
+	EIP721_NFT_TYPES,
 } from "./eip712"
 
 export type SimpleLazyNft<K extends keyof any> = Omit<LazyErc721, K> | Omit<LazyErc1155, K>
@@ -19,22 +19,28 @@ export async function signNft(ethereum: Ethereum, chainId: number, nft: SimpleLa
 		case "ERC721": {
 			const domain = createEIP712NftDomain(chainId, nft.contract, "ERC721")
 
-			const data = {
+			const data: TypedMessage<typeof EIP721_NFT_TYPES> = {
 				types: EIP721_NFT_TYPES,
 				domain,
 				primaryType: EIP721_NFT_TYPE,
-				message: { ...nft, tokenURI: nft.uri },
+				message: {
+					...nft,
+					tokenURI: nft.uri,
+				},
 			}
 			return toBinary(await signTypedData(ethereum, data))
 		}
 		case "ERC1155": {
 			const domain = createEIP712NftDomain(chainId, nft.contract, "ERC1155")
 
-			const data = {
+			const data: TypedMessage<typeof EIP1155_NFT_TYPES> = {
 				types: EIP1155_NFT_TYPES,
 				domain,
 				primaryType: EIP1155_NFT_TYPE,
-				message: { ...nft, tokenURI: nft.uri },
+				message: {
+					...nft,
+					tokenURI: nft.uri,
+				},
 			}
 			return toBinary(await signTypedData(ethereum, data))
 		}