+++ packages/protocol-ethereum-sdk/src/order/exchange-v1.test.ts
@@ -1,30 +1,32 @@
 import {
 	Configuration,
 	GatewayControllerApi,
-	NftCollectionControllerApi,
-	NftLazyMintControllerApi,
 	NftOwnershipControllerApi,
 	OrderControllerApi,
-	OrderForm
 } from "@rarible/protocol-api-client"
-import { toAddress, toBigNumber, toBinary } from "@rarible/types"
+import { toAddress, toBigNumber, toWord } from "@rarible/types"
 import { createE2eProvider } from "@rarible/ethereum-sdk-test-common"
-import { toBn } from "@rarible/utils/build/bn"
 import Web3 from "web3"
 import { Web3Ethereum } from "@rarible/web3-ethereum"
 import { CONFIGS } from "../config"
 import { retry } from "../common/retry"
+import { awaitAll } from "../common/await-all"
 import { send as sendTemplate } from "../common/send-transaction"
-import { signNft } from "../nft/sign-nft"
-import { mint as mintTemplate } from "../nft/mint"
 import { signOrder, SimpleOrder } from "./sign-order"
 import { fillOrderSendTx } from "./fill-order"
 import { getMakeFee } from "./get-make-fee"
+import { deployTestErc721 } from "./contracts/test/test-erc721"
 
 describe("test exchange v1 order", () => {
-	const { provider: provider1, wallet: wallet1 } = createE2eProvider()
-	const { provider: provider2, wallet: wallet2 } = createE2eProvider(
-		"ded057615d97f0f1c751ea2795bc4b03bbf44844c13ab4f5e6fd976506c276b9"
+	const {
+		provider: provider1,
+		wallet: wallet1,
+	} = createE2eProvider()
+	const {
+		provider: provider2,
+		wallet: wallet2,
+	} = createE2eProvider(
+		"ded057615d97f0f1c751ea2795bc4b03bbf44844c13ab4f5e6fd976506c276b9",
 	)
 	const web31 = new Web3(provider1)
 	const web32 = new Web3(provider2)
@@ -34,36 +36,34 @@ describe("test exchange v1 order", () => {
 	const configuration = new Configuration({ basePath: "https://ethereum-api-e2e.rarible.org" })
 	const orderApi = new OrderControllerApi(configuration)
 	const ownershipApi = new NftOwnershipControllerApi(configuration)
-	const nftCollectionApi = new NftCollectionControllerApi(configuration)
-	const nftLazyMintApi = new NftLazyMintControllerApi(configuration)
 	const gatewayApi = new GatewayControllerApi(configuration)
 	const send = sendTemplate.bind(null, gatewayApi)
-	const sign = signNft.bind(null, ethereum1, 17)
-	const mintHalfBind = mintTemplate.bind(null, ethereum1, send, sign, nftCollectionApi)
-	const mint = mintHalfBind.bind(null, nftLazyMintApi)
 
 	const seller = toAddress(wallet1.getAddressString())
 	const buyer = toAddress(wallet2.getAddressString())
 
-	const erc721ContractAddress = toAddress("0x22f8CE349A3338B15D7fEfc013FA7739F5ea2ff7")
+	const it = awaitAll({
+		testErc721: deployTestErc721(web31, "Test", "TST"),
+	})
+
+	const sign = signOrder.bind(null, ethereum1, {
+		chainId: 17,
+		exchange: CONFIGS.e2e.exchange,
+	})
+	const makeFee = getMakeFee.bind(null, { v2: 100 })
+	const fill = fillOrderSendTx
+		.bind(null, makeFee, ethereum2, send, CONFIGS.e2e.exchange)
+		.bind(null, orderApi)
 
 	test("", async () => {
-		const tokenId = await mint({
-			collection: {
-				id: erc721ContractAddress,
-				type: "ERC721",
-				supportsLazyMint: true,
-			},
-			uri: "uri",
-			creators: [{ account: toAddress(seller), value: 10000 }],
-			royalties: [],
-		})
+		const tokenId = toBigNumber("1")
+		await it.testErc721.methods.mint(seller, tokenId, "url").send({ from: seller })
 
-		let order: OrderForm = {
+		let order: SimpleOrder = {
 			make: {
 				assetType: {
 					assetClass: "ERC721",
-					contract: erc721ContractAddress,
+					contract: toAddress(it.testErc721.options.address),
 					tokenId: toBigNumber(tokenId),
 				},
 				value: toBigNumber("1"),
@@ -75,7 +75,7 @@ describe("test exchange v1 order", () => {
 				},
 				value: toBigNumber("100000"),
 			},
-			salt: toBigNumber("10") as any,
+			salt: toWord("0x000000000000000000000000000000000000000000000000000000000000000a"),
 			type: "RARIBLE_V1",
 			data: {
 				dataType: "LEGACY",
@@ -83,18 +83,13 @@ describe("test exchange v1 order", () => {
 			},
 		}
 
-		const leftSignature = await signOrder(
-			ethereum1,
-			{
-				chainId: 17,
-				exchange: CONFIGS.e2e.exchange,
-			},
-			orderFormToSimpleOrder(order)
-		)
+		await it.testErc721.methods.setApprovalForAll(CONFIGS.e2e.transferProxies.nft, true).send({ from: seller })
 
-		order = { ...order, signature: leftSignature }
-
-		await fillOrderSendTx(getMakeFee.bind(null, { v2: 100 }), ethereum2, send, CONFIGS.e2e.exchange, orderApi, order, {
+		const signedOrder = {
+			...order,
+			signature: await sign(order),
+		}
+		await fill(signedOrder, {
 			amount: 1,
 			payouts: [],
 			originFees: [],
@@ -102,16 +97,9 @@ describe("test exchange v1 order", () => {
 
 		await retry(10, async () => {
 			const ownership = await ownershipApi.getNftOwnershipById({
-				ownershipId: `${erc721ContractAddress}:${tokenId}:${buyer}`,
+				ownershipId: `${it.testErc721.options.address}:${tokenId}:${buyer}`,
 			})
 			expect(ownership.value).toBe("1")
 		})
 	}, 30000)
 })
-
-function orderFormToSimpleOrder(form: OrderForm): SimpleOrder {
-	return {
-		...form,
-		salt: toBinary(toBn(form.salt).toString(16)) as any,
-	}
-}