+++ packages/test-common/src/create-e2e-provider.ts
@@ -7,15 +7,19 @@ import { randomWord } from "@rarible/types"
 export function createE2eProvider(pk: string = randomWord()) {
 	const provider = new Web3ProviderEngine()
 	const wallet = new Wallet(Buffer.from(fixPK(pk), "hex"))
-	provider.addProvider(new TestSubprovider(wallet, { networkId: 17, chainId: 17 }))
+	provider.addProvider(new TestSubprovider(wallet, {
+		networkId: 17,
+		chainId: 17,
+	}))
 	provider.addProvider(new RpcSubprovider({ rpcUrl: "https://node-e2e.rarible.com" }))
 
-	beforeAll(async () => {
-		await provider.start()
+	beforeAll(() => {
+		provider.start()
 	})
 
-	afterAll(async () => {
-		await provider.stop()
+	afterAll((done) => {
+		provider.stop()
+		done()
 	})
 
 	return {