+++ packages/web3-ethereum/src/index.ts
@@ -5,7 +5,7 @@ import type {
 	EthereumContract,
 	EthereumFunctionCall,
 	EthereumSendOptions,
-	EthereumTransaction
+	EthereumTransaction,
 } from "@rarible/ethereum-provider"
 import { Address, Binary, toAddress, toBinary, toWord, Word } from "@rarible/types"
 import { backOff } from "exponential-backoff"