+++ packages/protocol-ethereum-sdk/CHANGELOG.md
@@ -3,6 +3,90 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.15](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.14...v0.1.15) (2021-09-10)
+
+
+### Bug Fixes
+
+* disable test ([62bebc0](https://github.com/rariblecom/protocol-ethereum-sdk/commit/62bebc0388a63a269518a087b85447a0d3394a90))
+
+
+
+
+
+## [0.1.14](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.13...v0.1.14) (2021-09-10)
+
+
+### Bug Fixes
+
+* disable test ([94454e6](https://github.com/rariblecom/protocol-ethereum-sdk/commit/94454e6b0948ebc93756e06758c1f374ebf09fba))
+
+
+
+
+
+## [0.1.13](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.12...v0.1.13) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/protocol-ethereum-sdk
+
+
+
+
+
+## [0.1.12](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.11...v0.1.12) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/protocol-ethereum-sdk
+
+
+
+
+
+## [0.1.11](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.10...v0.1.11) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/protocol-ethereum-sdk
+
+
+
+
+
+## [0.1.10](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.9...v0.1.10) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/protocol-ethereum-sdk
+
+
+
+
+
+## [0.1.9](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.8...v0.1.9) (2021-09-10)
+
+
+### Bug Fixes
+
+* fixed dependencies for sdk ([46176a8](https://github.com/rariblecom/protocol-ethereum-sdk/commit/46176a800ddba677881101d786ca002156de9486))
+
+
+
+
+
+## [0.1.8](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.7...v0.1.8) (2021-09-10)
+
+
+### Bug Fixes
+
+* upgraded api-client, improved types ([129d179](https://github.com/rariblecom/protocol-ethereum-sdk/commit/129d1794379f813d199e403b06a365f7c52eba44))
+
+
+
+
+
+## [0.1.7](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.6...v0.1.7) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/protocol-ethereum-sdk
+
+
+
+
+
 ## [0.1.6](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.5...v0.1.6) (2021-09-03)