+++ packages/test-common/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.11](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.10...v0.1.11) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/ethereum-sdk-test-common
+
+
+
+
+
+## [0.1.10](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.9...v0.1.10) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/ethereum-sdk-test-common
+
+
+
+
+
+## [0.1.7](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.6...v0.1.7) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/ethereum-sdk-test-common
+
+
+
+
+
 ## [0.1.6](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.5...v0.1.6) (2021-09-03)