+++ packages/web3-ethereum/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.13](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.12...v0.1.13) (2021-09-10)
+
+
+### Bug Fixes
+
+* improve web3-ethereum deps ([f98bae9](https://github.com/rariblecom/protocol-ethereum-sdk/commit/f98bae92784afbaa39e9c51a2cd889825007a469))
+
+
+
+
+
+## [0.1.12](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.11...v0.1.12) (2021-09-10)
+
+
+### Bug Fixes
+
+* improve package.json ([49bfd86](https://github.com/rariblecom/protocol-ethereum-sdk/commit/49bfd86f33aabdba1838e843f253b2cdaf6b0cb7))
+
+
+
+
+
+## [0.1.11](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.10...v0.1.11) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/web3-ethereum
+
+
+
+
+
+## [0.1.10](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.9...v0.1.10) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/web3-ethereum
+
+
+
+
+
+## [0.1.8](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.7...v0.1.8) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/web3-ethereum
+
+
+
+
+
+## [0.1.7](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.6...v0.1.7) (2021-09-10)
+
+**Note:** Version bump only for package @rarible/web3-ethereum
+
+
+
+
+
 ## [0.1.6](https://github.com/rariblecom/protocol-ethereum-sdk/compare/v0.1.5...v0.1.6) (2021-09-03)